moved Transform to FamilyInstance + added check to see if the element has already been transformed
